requirements.txt
alembic==1.10.2
anyio==3.6.2
fastapi==0.94.1
greenlet==2.0.2
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
pydantic==1.10.6
sniffio==1.3.0
SQLAlchemy==2.0.6
SQLAlchemy-Utils==0.38.3
starlette==0.26.1
starlette-context==0.3.4
typing_extensions==4.5.0
psycopg2-binary==2.9.3
python-multipart==0.0.5
PyJWT==2.6.0
uvicorn==0.18.2
python-dotenv==0.20.0